\name{ivFixed.formula}
\alias{ivFixed.formula}
\title{formula}
\usage{
\method{ivFixed}{formula}(formula, data = list(), n, t, ...)
}
\arguments{
  \item{formula}{PIB~INF+TIR|Cap+m2r}

  \item{data}{the dataframe}

  \item{n}{the number of section}

  \item{t}{the time per section}

  \item{...}{not used}
}
\description{
formula
}

